refactor libs
