+++ src/components/cart/EmptyCart.tsx
@@ -7,7 +7,7 @@ import { useRouter } from "next/navigation";
 
 import { Input } from "@/components/ui/input";
 
-import { handleSearch } from "@/lib/utils";
+import { handleSearch } from "@/libs/navigationUtils";
 
 const montserrat = Montserrat({
   subsets: ["cyrillic", "latin", "vietnamese"],
