+++ src/components/header/DesktopHeader.tsx
@@ -12,8 +12,8 @@ import { Input } from "@/components/ui/input";
 
 import { useCart } from "@/hooks/useCart";
 
-import { HASH_ROUTES, ROUTES } from "@/constants/routes";
-import { handleNavigation, handleSearch } from "@/lib/utils";
+import { HASH_ROUTES } from "@/constants/routes";
+import { handleNavigation, handleSearch } from "@/libs/navigationUtils";
 
 const montserrat = Montserrat({
   subsets: ["cyrillic", "latin", "vietnamese"],
@@ -73,6 +73,7 @@ const DesktopHeader: React.FC<DesktopHeaderProps> = ({
               handleNavigation(
                 HASH_ROUTES.COLLECTION,
                 hasSections,
+                undefined,
                 router,
                 collectionRef,
               );
@@ -88,6 +89,7 @@ const DesktopHeader: React.FC<DesktopHeaderProps> = ({
               handleNavigation(
                 HASH_ROUTES.PRODUCTS,
                 hasSections,
+                undefined,
                 router,
                 productsRef,
               );
@@ -103,6 +105,7 @@ const DesktopHeader: React.FC<DesktopHeaderProps> = ({
               handleNavigation(
                 HASH_ROUTES.ABOUT,
                 hasSections,
+                hasFooter,
                 router,
                 aboutRef,
               );
