+++ src/components/ui/carousel.tsx
@@ -9,7 +9,7 @@ import useEmblaCarousel, {
 
 import { Button } from "@/components/ui/button";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 type CarouselApi = UseEmblaCarouselType[1];
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>;
