+++ src/components/ui/product-carousel.tsx
@@ -14,7 +14,7 @@ import {
   CarouselPrevious,
 } from "@/components/ui/carousel";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 interface ProductCarouselProps {
   value?: string | null;
