+++ src/components/ui/tooltip.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 const TooltipProvider = TooltipPrimitive.Provider;
