+++ src/libs/productUtils.ts
